requirements.txt
# requirements.txt — Python dependencies for PC Automation Bot
#
# Install with:  pip install -r requirements.txt

# Telegram Bot API wrapper (v20+ uses asyncio natively)
python-telegram-bot==22.6

# python-telegram-bot pulls these in, but pin them for stability:
httpx>=0.27.0
anyio>=4.4.0

# Screenshot capture
mss>=9.0.1
Pillow>=10.0.0
